Organize msg structs
